Fix migrations: V3 keep server version, V4 autopilot interval

+++ src/main/java/ru/rentoptima/repository/BookingRepository.java
@@ -55,5 +55,5 @@ public interface BookingRepository extends JpaRepository<Booking, Long> {
 
     @Modifying
     @Query("DELETE FROM Booking b WHERE b.tenant.id = :tenantId AND b.property.id = :propertyId")
-    int deleteByTenantIdAndPropertyId(Long tenantId, Long propertyId);
+    long deleteByTenantIdAndPropertyId(Long tenantId, Long propertyId);
 }
